refactor: split agent_helpers into status-checks, runtime-wiring, embedder modules
