+++ CLAUDE.md
@@ -89,7 +89,7 @@ PYTHONPATH=. uv run zensical build                  # docs
 - Self-extending toolkit autonomous detection wires in `wire_toolsmith` AFTER `install_dynamic_tool_layer` (gated on `tool_creation_enabled` + provider + connected persistence): `install_capability_gap_sink(runtime.service)` routes every `capability_gap` MCP envelope into the gap store (fire-and-forget; write failure logs via `safe_error_description`, never an unhandled task-exception traceback), and a `ToolsmithCycleScheduler` (cadence `toolsmith.cycle_interval_seconds`, floor 60s) drives `run_cycle` periodically with a `meta.toolsmith_cycle_paused` kill-switch (re-read each tick, fail-safe to enabled). The `ToolsmithStateSlice` carries `service` + `cycle_scheduler` as a both-or-neither paired invariant; `lifecycle_runner_shutdown` stops the scheduler. The cycle only PROPOSES (human approval still gates apply).
 - Red-team completion gate is stakes-gated: `_wire_red_team_completion` threads `stakes_routing.red_team_min_stakes` (default HIGH) onto `ReviewGateService` via `set_red_team_min_stakes`, so the gate fires only for `task.stakes >= red_team_min_stakes`; a below-threshold completion logs `RED_TEAM_GATE_SKIPPED` and `dispatch_completion` runs it INLINE (a missing task also runs inline so `complete_review` raises `TaskNotFoundError` synchronously) rather than deferring a no-op to the background.
 - Runtime services: `synthorg.workers.runtime_builder.build_runtime_services` selects behind ONE provider-present switch and returns a `RuntimeServices` pair (worker execution service + multi-agent coordinator) built from a SINGLE shared boot `AgentEngine`: `AgentEngineExecutionService` + a `build_coordinator(...)` coordinator with a provider, `NoProviderExecutionService` + `None` coordinator as the empty-company backstop. The `_install_runtime_services` boot hook installs both via the `AppState.worker_execution_service` and `AppState.coordinator` seams; it is appended FIRST after the persistence/SettingsService hooks so the once-only `set_worker_execution_service` and if-absent `set_coordinator_if_absent` seams cannot lose the race with the worker property's lazy `LifecycleAdvancingExecutionService` default. Empty-company rejects task creation at the controller (`AgentRuntimeNotConfiguredError`, 4014) and `/coordinate` honestly 503s (no coordinator). `swap_worker_execution_service` / `swap_coordinator` / `swap_provider_registry` hold a lock (synchronised against lazy reads).
-- Setup completion: `post_setup_reinit()` (provider reload, agent bootstrap, AND runtime-services rebuild + dual hot-swap of the worker execution service and coordinator, defined in `src/synthorg/api/controllers/setup/agent_helpers.py`) propagates failures, and `settings_svc.set("api", "setup_complete", "true")` only runs if reinit returns clean. The whole check/validate/reinit/persist sequence is serialised under `COMPLETE_LOCK` in the same module so two concurrent `/setup/complete` requests cannot race on the flag write. A half-configured runtime presenting itself as "complete" is worse than a clear error the operator can retry after fixing the underlying provider config.
+- Setup completion: `post_setup_reinit()` (provider reload, agent bootstrap, AND runtime-services rebuild + dual hot-swap of the worker execution service and coordinator, defined in `src/synthorg/api/controllers/setup/_runtime_wiring.py`) propagates failures, and `settings_svc.set("api", "setup_complete", "true")` only runs if reinit returns clean. The whole check/validate/reinit/persist sequence is serialised under `COMPLETE_LOCK` (also in `_runtime_wiring.py`) so two concurrent `/setup/complete` requests cannot race on the flag write. A half-configured runtime presenting itself as "complete" is worse than a clear error the operator can retry after fixing the underlying provider config.
 
 ## MCP / Telemetry / Resilience
