+++ src/synthorg/api/controllers/setup/__init__.py
@@ -5,8 +5,10 @@ controller modules -- ``status`` (status probe + template listing),
 ``company`` (company creation), ``agents`` (agent CRUD), ``locales``
 (name-locale configuration), and ``completion`` (the completion flow
 plus its prerequisite-validation helpers) -- alongside the shared
-``agent_helpers`` (agent-bootstrap / locks) and ``company_helpers``
-(company-metadata) helper modules.
+agent-side helper modules ``_status_checks`` (read-only probes),
+``_runtime_wiring`` (provider reload / hot-swap / setup locks), and
+``_embedder_setup`` (template-agent creation / embedder selection), plus
+``company_helpers`` (company-metadata).
 
 Direct imports only:
 ``from synthorg.api.controllers.setup.<module> import ...``.
